implemented removeEdge and fixed addEdge
A wrongful assumption resulted in invalid reads when some test cases where ran. The assumption was that if a vertex exists in the graph, then it must have an associated adjacency list. This does not hold when a vertex is added as the "to vertex" of a directed edge and then later added in an edge as a "from vertex"